Feat: #108 UserRepository 인터페이스 패턴 리팩토링
- UserRepository를 인터페이스로 분리
- UserRepositoryImpl 구현체 생성
- 사용자 정보 조회/수정/로그아웃 기능 인터페이스 적용
