+++ app/src/main/java/com/konkuk/medicarecall/data/repository/UserRepository.kt
@@ -1,47 +1,9 @@
 package com.konkuk.medicarecall.data.repository
 
-import android.util.Log
-import com.konkuk.medicarecall.data.api.SettingService
 import com.konkuk.medicarecall.data.dto.response.MyInfoResponseDto
-import retrofit2.HttpException
-import javax.inject.Inject
 
-class UserRepository @Inject constructor(
-    private val settingService: SettingService,
-    private val tokenStore: DataStoreRepository
-) {
-    suspend fun getMyInfo() = runCatching {
-        val response = settingService.getMyInfo()
-        if (response.isSuccessful) {
-            response.body() ?: throw IllegalStateException("Response body is null")
-        } else {
-            val errorBody = response.errorBody()?.string() ?: "Unknown error"
-            throw HttpException(response)
-        }
-    }
-
-    suspend fun updateMyInfo(userUpdateRequestDto: MyInfoResponseDto) = runCatching {
-        Log.d("UserRepository", "updateMyInfo() 진입: $userUpdateRequestDto")
-        val response = settingService.updateMyInfo(userUpdateRequestDto)
-        if (response.isSuccessful) {
-            response.body() ?: throw IllegalStateException("Response body is null")
-        } else {
-            throw HttpException(response)
-        }
-    }
-
-    suspend fun logout(): Result<Unit> {
-        val result = runCatching {
-            val refresh = tokenStore.getRefreshToken() ?: error("Refresh token is null")
-            val response = settingService.logout("Bearer $refresh")
-            if (!response.isSuccessful) {
-                val errorBody = response.errorBody()?.string() ?: "Unknown error"
-                throw HttpException(response)
-            }
-            Unit
-        }
-        // 성공/실패와 무관하게 로컬 토큰 제거(보안/UX 측면에서 권장)
-        tokenStore.clearTokens()
-        return result
-    }
+interface UserRepository {
+    suspend fun getMyInfo(): Result<MyInfoResponseDto>
+    suspend fun updateMyInfo(userUpdateRequestDto: MyInfoResponseDto): Result<MyInfoResponseDto>
+    suspend fun logout(): Result<Unit>
 }
